Update child theme files.
- Move 'enqueue_child_styles' from functions.php to enqueue-assets.php file. 
- Call enqueue-assets.php with include in functions.php.
- Reorganize helper functions in functions.php.
